Use same API query-parsing code for all endpoints
Move our custom query-decoding and query-encoding process to a separate
component and use it on all API endpoints. Although this query-parsing
code is actually only used on /unstable/sentences endpoint so far,
this commit frees SentencesController from these specifics and ensures
query-parsing is consistent over the whole API.
