+++ src/Controller/VHosts/Api/ApiController.php
@@ -158,6 +158,7 @@ class ApiController extends Controller
         $this->loadComponent('RequestHandler', [
             'viewClassMap' => ['json' => 'Api']
         ]);
+        $this->loadComponent('Api');
     }
 
     public function beforeFilter(Event $event)
